requirements.txt
websockets==10.0
python-dotenv